fix: Arreglar problema con los tipos de datos en las entidades
Se arreglaron unos problemas con los tipos de datos en las entidades entre Double y BigDecimal

+++ kinalcoffeeshop/src/main/java/org/asco_devs/kinalcoffeeshop/dominio/dto/ModPagoDto.java
@@ -3,6 +3,6 @@ package org.asco_devs.kinalcoffeeshop.dominio.dto;
 public record ModPagoDto(
         String amount,
         String type,
-        Long orderId
+        Integer orderId
 ) {
 }
